requirements.txt
# GitHub API
PyGithub>=1.55

# Data Processing
pandas>=2.1
numpy>=1.26

# NLP & Embeddings
sentence-transformers>=2.2
torch>=2.0

# Semantic Search
faiss-cpu>=1.7

# ML Models
xgboost>=2.0
scikit-learn>=1.3

# Graph Analysis
networkx>=3.0

# Text Ranking
rank-bm25>=0.2

# API Server
Flask>=2.3
python-dotenv>=1.0

# Utilities
tqdm>=4.60
python-dateutil>=2.8
requests>=2.30

# Testing
pytest>=7.0
